requirements.txt
nibabel
numpy
matplotlib
tqdm
matplotlib
scipy
ipykernel
torch
torchvision
Pillow
image-utils
dominate
bs4

